refactor(css-injector): merge Css Injectors
